backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
beautifulsoup4==4.12.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.38
pytz==2023.3
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
zope.event==4.6
zope.interface==6.0